style: int

+++ practice/6_29/Q2.py
@@ -1,8 +1,8 @@
 # zerojudge: https://zerojudge.tw/ShowProblem?problemid=a216
 def f(n):
-    return (n**2 + n) / 2
+    return int((n**2 + n) / 2)
 def g(n):
-    return n * (n**2 + 3*n + 2) / 6
+    return int(n * (n**2 + 3*n + 2) / 6)
 
 def main():
     while True:
@@ -10,5 +10,5 @@ def main():
             n = int(input())
         except:
             break
-        print(int(f(n)), int(g(n)))
+        print(f(n), g(n))
 main()
